Added trailing div to prevent jumping when help is turned off
Some of the help texts are big, and when you turn them off, the view would scroll down so the last thing is at the bottom of the view port, which was rather distracting and uncomfortable. This addition allows scrolling past the last viewable thing (90vh means 90% of the height of the users browser view port), effectively preventing this scrolling effect when help texts are hidden.